requirements.txt
pandas~=2.3.0
fastapi~=0.115.13
uvicorn~=0.34.3
python-dotenv~=1.1.0
motor~=3.7.1
asyncpg~=0.30.0
pydantic~=2.11.7
python-multipart
sqlalchemy
pymongo